Revert "Release v0.0.1 (#4)"
This reverts commit b9e9abd88bbb351eb9d55bb4625ac36713d96031.
