--- .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,31 +1,0 @@
----
-name: Bug Report
-about: Report a bug
-title: "[Bug]: <short descriptive title>"
-labels: bug
-assignees: ''
----
-
-## Summary
-
-<!-- Briefly describe the bug. Include the observed behavior. -->
-
-## Steps to Reproduce
-
-<!-- List clear, minimal steps to reproduce the bug. Example:
-1. Go to '...'
-2. Click on '...'
-3. Observe the error
--->
-
-## Expected Behavior
-
-<!-- Describe what you expected to happen. -->
-
-## Actual Behavior
-
-<!-- Describe what actually happened, including error messages if any. -->
-
-## Additional Notes
-
-<!-- Optional: Screenshots, logs, or other context to help reproduce/fix the bug. -->
