--- .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,28 +1,0 @@
----
-name: Feature Request
-about: Propose a new feature or enhancement for the project
-title: "[<FeatureName>]: <short descriptive title>"
-labels: enhancement
-assignees: ''
----
-
-<!-- Don't forget to replace <FeatureName> by a short name with which you'd name the feature, and <short descriptive title> by an appropriate short title -->
-## Summary
-
-<!-- Make a brief but concise description of the feature/enhancement. Don't exceed two sentences -->
-
-## Why
-
-<!-- Explain the motivation behind this change. Why is it needed? -->
-
-## What Remains To Do
-
-<!-- List the TODOs required for the issue completion. -->
-<!-- Use this format:
-- [ ] Task description ending with a period.
-- [x] Same formatting, but the task is done (x between the brackets).
--->
-
-## Additional Notes
-
-<!-- Any extra details, future improvements, or related discussions. -->
