--- .github/ISSUE_TEMPLATE/question.md
@@ -1,23 +1,0 @@
----
-name: Question
-about: Ask a question or request clarification
-title: "[Question]: <short descriptive title>"
-labels: question
-assignees: ''
----
-
-## Summary
-
-<!-- Briefly describe your question. -->
-
-## Context
-
-<!-- Provide background or context to help others understand your question. -->
-
-## What You Want to Achieve
-
-<!-- Describe what outcome or clarification you are looking for. -->
-
-## Additional Notes
-
-<!-- Optional: Any extra details, links, or references. -->
