+++ Cargo.toml
@@ -4,6 +4,3 @@ version = "0.0.1"
 edition = "2024"
 
 [dependencies]
-
-[[example]]
-name = "schedules"
